requirements.txt
python-evtx
networkx
pyvis
typer
lxml
matplotlib
pytest
